json export of pages and code reorganization
