+++ CHANGELOG.md
@@ -3,3 +3,4 @@
 - Rest API
 - 4 types of audit
 - User, contract and project management
+- JSON export of audits and pages
