+++ tanaguru-engine/tanaguru-api/pom.xml
@@ -151,5 +151,10 @@
             <scope>test</scope>
         </dependency>
 
+		<dependency>
+		    <groupId>org.json</groupId>
+		    <artifactId>json</artifactId>
+		    <version>20180130</version>
+		</dependency>
     </dependencies>
 </project>
