+++ tanaguru-engine/tanaguru-api/src/main/java/com/tanaguru/service/PageService.java
@@ -1,5 +1,7 @@
 package com.tanaguru.service;
 
+import org.json.JSONObject;
+
 import com.tanaguru.domain.entity.audit.Audit;
 import com.tanaguru.domain.entity.audit.Page;
 
@@ -15,4 +17,11 @@ public interface PageService {
      * @param audit The given @see Audit
      */
     void deletePageByAudit(Audit audit);
+    
+    /**
+     * Return a json object with page information
+     * @param page the given @see Page
+     * @return json object
+     */
+    JSONObject exportPage(Page page);
 }
