+++ tanaguru-rest/src/main/java/com/tanaguru/controller/AuditController.java
@@ -1,9 +1,5 @@
 package com.tanaguru.controller;
 
-import com.fasterxml.jackson.core.JsonProcessingException;
-import com.fasterxml.jackson.databind.ObjectMapper;
-import org.json.JSONArray;
-import org.json.JSONException;
 import org.json.JSONObject;
 
 import com.tanaguru.domain.constant.EAuditParameter;
@@ -11,19 +7,9 @@ import com.tanaguru.domain.constant.EAuditType;
 import com.tanaguru.domain.dto.AuditCommandDTO;
 import com.tanaguru.domain.dto.DemoCommandDTO;
 import com.tanaguru.domain.entity.audit.Audit;
-import com.tanaguru.domain.entity.audit.AuditLog;
-import com.tanaguru.domain.entity.audit.AuditReference;
-import com.tanaguru.domain.entity.audit.AuditScheduler;
-import com.tanaguru.domain.entity.audit.Page;
-import com.tanaguru.domain.entity.audit.TanaguruTest;
 import com.tanaguru.domain.entity.audit.TestHierarchy;
-import com.tanaguru.domain.entity.audit.parameter.AuditAuditParameterValue;
 import com.tanaguru.domain.entity.membership.Act;
 import com.tanaguru.domain.entity.membership.project.Project;
-import com.tanaguru.domain.entity.pageresult.ElementResult;
-import com.tanaguru.domain.entity.pageresult.StatusResult;
-import com.tanaguru.domain.entity.pageresult.TestHierarchyResult;
-import com.tanaguru.domain.entity.pageresult.TestResult;
 import com.tanaguru.domain.exception.ForbiddenException;
 import com.tanaguru.domain.exception.InvalidEntityException;
 import com.tanaguru.factory.AuditFactory;
@@ -39,14 +25,11 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.core.io.ByteArrayResource;
-import javax.persistence.EntityNotFoundException;
 import org.springframework.core.io.Resource;
+import javax.persistence.EntityNotFoundException;
 import javax.validation.Valid;
 
-import java.text.DateFormat;
-import java.text.SimpleDateFormat;
 import java.util.*;
-import java.util.stream.Collectors;
 
 /**
  * @author rcharre
@@ -62,15 +45,6 @@ public class AuditController {
     private final ProjectRepository projectRepository;
     private final ActRepository actRepository;
     private final TestHierarchyRepository testHierarchyRepository;
-    private final TestHierarchyResultRepository testHierarchyResultRepository;
-    private final AuditLogRepository auditLogRepository;
-    private final AuditReferenceRepository auditReferenceRepository;
-    private final AuditAuditParameterValueRepository auditAuditParameterValueRepository;
-    private final AuditSchedulerRepository auditSchedulerRepository;
-    private final StatusResultRepository statusResultRepository;
-    private final TestResultRepository testResultRepository;
-    private final ElementResultRepository elementResultRepository;
-    private final PageRepository pageRepository;
     
     @Autowired
     public AuditController(
@@ -79,16 +53,7 @@ public class AuditController {
             AuditRunnerService auditRunnerService,
             ProjectRepository projectRepository,
             ActRepository actRepository,
-            TestHierarchyRepository testHierarchyRepository,
-            AuditLogRepository auditLogRepository,
-            AuditReferenceRepository auditReferenceRepository,
-            AuditAuditParameterValueRepository auditAuditParameterValueRepository,
-            AuditSchedulerRepository auditSchedulerRepository,
-            StatusResultRepository statusResultRepository,
-            TestResultRepository testResultRepository,
-            ElementResultRepository elementResultRepository,
-            TestHierarchyResultRepository testHierarchyResultRepository,
-            PageRepository pageRepository) {
+            TestHierarchyRepository testHierarchyRepository) {
 
         this.auditRepository = auditRepository;
         this.auditService = auditService;
@@ -97,15 +62,6 @@ public class AuditController {
         this.projectRepository = projectRepository;
         this.actRepository = actRepository;
         this.testHierarchyRepository = testHierarchyRepository;
-        this.auditLogRepository = auditLogRepository;
-        this.auditReferenceRepository = auditReferenceRepository;
-        this.auditAuditParameterValueRepository = auditAuditParameterValueRepository;
-        this.auditSchedulerRepository = auditSchedulerRepository;
-        this.statusResultRepository = statusResultRepository;
-        this.testResultRepository = testResultRepository;
-        this.elementResultRepository = elementResultRepository;
-        this.testHierarchyResultRepository = testHierarchyResultRepository;
-        this.pageRepository = pageRepository;
     }
 
     /**
@@ -139,7 +95,7 @@ public class AuditController {
      * Get a json file with the audit information
      * @param id The id of the @see Audit
      * @param shareCode the share code of the @see Audit
-     * @return @see Audit
+     * @return resource json
      */
     @ApiOperation(
             value = "Get a json file with the audit information",
@@ -157,7 +113,7 @@ public class AuditController {
             @PathVariable long id,
             @ApiParam(required = false) @PathVariable(required = false) String shareCode) {
         Audit audit = auditRepository.findById(id).orElseThrow(EntityNotFoundException::new);
-        JSONObject jsonFinalObject = exportAudit(audit);
+        JSONObject jsonFinalObject = auditService.exportAudit(audit);
         byte[] buf = jsonFinalObject.toString().getBytes();              
         HttpHeaders header = setUpHeaders(audit.getName());
         return ResponseEntity
@@ -355,204 +311,10 @@ public class AuditController {
      */
     private HttpHeaders setUpHeaders(String filename) {
         HttpHeaders header = new HttpHeaders();
-        header.add("Content-Disposition", "attachment; filename=\""+filename+"\"");
-        header.add("Cache-Control", "no-cache, no-store, must-revalidate");
+        header.add("Content-Disposition", "attachment; filename=\""+filename+".json\"");
+        header.add("Cache-Control", "no-store");
         header.add("Pragma", "no-cache");
-        header.add("Expires", "0");
         return header;
     }
-    
-    /**
-     * Return information of the audit as a json object
-     * @param audit the audit to export
-     * @return json object of the audit
-     */
-    private JSONObject exportAudit(Audit audit) {
-        JSONObject jsonFinalObject = new JSONObject();
-        ObjectMapper mapper = new ObjectMapper();
-        DateFormat df = new SimpleDateFormat("yyyy-MM-dd HH:mm:ss");
-        mapper.setDateFormat(df);
-        try {
-            addLogs(jsonFinalObject,audit, mapper);
-            addAct(jsonFinalObject,audit, mapper);
-            addScheduler(jsonFinalObject,audit, mapper);
-            addParametersValues(jsonFinalObject,audit, mapper);
-            addReferenceAndResults(jsonFinalObject,audit, mapper);    
-        } catch (JSONException | JsonProcessingException e) {
-            e.printStackTrace();
-        }
-        return jsonFinalObject;
-    }
-    
-    /**
-     * Add logs information of the audit to the json object
-     * @param jsonFinalObject the json object
-     * @param audit the audit
-     * @param mapper object mapper to json
-     * @throws JSONException
-     * @throws JsonProcessingException
-     */
-    private void addLogs(JSONObject jsonFinalObject, Audit audit, ObjectMapper mapper) throws JsonProcessingException {
-        Collection<AuditLog> auditLogs =  auditLogRepository.findAllByAudit(audit);
-        String auditLog = "auditLogs";
-        if(!auditLogs.isEmpty()) {
-            JSONObject jsonAuditLogsObject = new JSONObject();
-            for(AuditLog log : auditLogs) {
-                jsonAuditLogsObject.append(auditLog, new JSONObject(mapper.writeValueAsString(log)));   
-            }
-            jsonFinalObject.put(auditLog, jsonAuditLogsObject.get(auditLog));
-        }
-    }
-    
-    /**
-     * Add act information of the audit to the json object
-     * @param jsonFinalObject the json object
-     * @param audit the audit
-     * @param mapper object mapper to json
-     * @throws JSONException
-     * @throws JsonProcessingException
-     */
-    private void addAct(JSONObject jsonFinalObject, Audit audit, ObjectMapper mapper) throws JsonProcessingException {
-        Optional<Act> act = actRepository.findByAudit(audit);
-        if(!act.isEmpty()) {
-            JSONObject jsonActObject = new JSONObject(mapper.writeValueAsString(act.get()));
-            jsonFinalObject.put("act", jsonActObject);
-        }
-    }
-    
-    /**
-     * Add parameters information of the audit to the json object
-     * @param jsonFinalObject the json object
-     * @param audit the audit
-     * @param mapper object mapper to json
-     * @throws JSONException
-     * @throws JsonProcessingException
-     */
-    private void addParametersValues(JSONObject jsonFinalObject, Audit audit, ObjectMapper mapper) throws JsonProcessingException {
-        Collection<AuditAuditParameterValue> auditAuditParameterValues = auditAuditParameterValueRepository.findAllByAudit(audit);
-        if(!auditAuditParameterValues.isEmpty()) {
-            JSONObject jsonAuditParameterObject = new JSONObject();
-            for(AuditAuditParameterValue auditParameterValue : auditAuditParameterValues) {
-                jsonAuditParameterObject.append("parameters", new JSONObject(mapper.writeValueAsString(auditParameterValue.getAuditParameterValue())));
-            }
-            jsonFinalObject.put("auditParametersValues", jsonAuditParameterObject.get("parameters"));
-        }
-    }
-    
-    /**
-     * Add scheduler information of the audit to the json object
-     * @param jsonFinalObject the json object
-     * @param audit the audit
-     * @param mapper object mapper to json
-     * @throws JSONException
-     * @throws JsonProcessingException
-     */
-    private void addScheduler(JSONObject jsonFinalObject, Audit audit, ObjectMapper mapper) throws JsonProcessingException {
-        Optional<AuditScheduler> auditScheduler = auditSchedulerRepository.findByAudit(audit);
-        if(!auditScheduler.isEmpty()) {
-            jsonFinalObject.put("auditScheduler", mapper.writeValueAsString(auditScheduler.get()));
-        }
-    }
-    
-    /**
-     * Add reference and results information of the audit to the json object
-     * @param jsonFinalObject the json object
-     * @param audit the audit
-     * @param mapper object mapper to json
-     * @throws JSONException
-     * @throws JsonProcessingException
-     */
-    private void addReferenceAndResults(JSONObject jsonFinalObject, Audit audit, ObjectMapper mapper) throws JsonProcessingException {
-        Collection<AuditReference> auditReferences = auditReferenceRepository.findAllByAudit(audit);
-        for(AuditReference ref : auditReferences) { //for each repository
-            JSONObject jsonAuditReferenceObject = new JSONObject();
-            jsonAuditReferenceObject.put("referenceTestHierarchyName", ref.getTestHierarchy().getName());
-            jsonAuditReferenceObject.put("referenceTestHierarchyCode", ref.getTestHierarchy().getCode());
-            jsonAuditReferenceObject.put("referenceTestHierarchyUrls", ref.getTestHierarchy().getUrls());
-            jsonAuditReferenceObject.put("referenceTestHierarchyIsMain", ref.isMain());
-            jsonAuditReferenceObject.put("referenceTestHierarchyId", ref.getTestHierarchy().getId());
-            Collection<StatusResult> statusResults = statusResultRepository.findAllByReferenceAndPage_Audit(ref.getTestHierarchy(), audit);
-            for(StatusResult statusResult : statusResults) { //for each page of the audit
-                JSONObject jsonStatusResultsObject = new JSONObject();
-                JSONObject jsonPageStatusResultsObject = new JSONObject(mapper.writeValueAsString(statusResult));
-                
-                Collection<TestResult> testsResults = testResultRepository.findDistinctByPageAndTanaguruTest_TestHierarchies_Reference(
-                        statusResult.getPage(),
-                        ref.getTestHierarchy());
-                for(TestResult testResult : testsResults) { //for each test
-                    TanaguruTest tanaguruTest = testResult.getTanaguruTest();                       
-                    JSONObject oneTestResult = new JSONObject(mapper.writeValueAsString(testResult));
-                    JSONArray elementResults = oneTestResult.getJSONArray("elementResults");
-                    List<Long> longs = elementResults.toList().stream()
-                            .map(object -> Long.parseLong(String.valueOf(object)))
-                            .collect(Collectors.toList());
-                    Collection<ElementResult> elements = elementResultRepository.findAllByIdIn(longs);
-                    oneTestResult.put("elementResults", new JSONArray(mapper.writeValueAsString(elements)));
-
-                    JSONObject tanaguruTestJson = new JSONObject();
-                    tanaguruTestJson.put("name", tanaguruTest.getName());
-                    tanaguruTestJson.put("description", tanaguruTest.getDescription());
-                    for(TestHierarchy th : tanaguruTest.getTestHierarchies()) { //for each test hierarchy corresponding to the tanaguru test
-                        JSONObject testHierarchyInfos = new JSONObject();
-                        testHierarchyInfos.put("name", th.getName());
-                        testHierarchyInfos.put("code", th.getCode());
-                        testHierarchyInfos.put("id", th.getId());
-                        testHierarchyInfos.put("urls", th.getUrls());
-                        tanaguruTestJson.append("testHierarchy", testHierarchyInfos);
-                    }
-                    oneTestResult.put("tanaguruTest", tanaguruTestJson);
-                    jsonPageStatusResultsObject.append("testsResults", oneTestResult);
-                }
-                //for each page of the audit add all the test hierarchy and their results (of the repository)
-                addTestHierarchy(jsonPageStatusResultsObject, statusResult.getPage(), ref.getTestHierarchy().getId());
-                jsonStatusResultsObject.put("pageName", statusResult.getPage().getName());
-                jsonStatusResultsObject.put("pageId", statusResult.getPage().getId());
-                jsonStatusResultsObject.put("pageResults", jsonPageStatusResultsObject);
-                jsonAuditReferenceObject.append("auditResults", jsonStatusResultsObject);
-            }
-            jsonFinalObject.append("auditReference", jsonAuditReferenceObject);
-        }
-    }    
-    
-    /**
-     * Add test hierarchy results of the page to the json object
-     * @param pageResultObject intermediate json object (page results)
-     * @param page the page
-     * @param referenceTestHierarchyId the reference test hierarchy id
-     */
-    private void addTestHierarchy(JSONObject pageResultObject, Page page, long referenceTestHierarchyId) {
-        Collection<TestHierarchy> testHierarchies = testHierarchyRepository.findAllByReferenceId(referenceTestHierarchyId);
-        for(TestHierarchy th : testHierarchies) { //for each test hierarchy of the repository
-            Optional<TestHierarchyResult> testHierarchyResult = testHierarchyResultRepository.findByTestHierarchyAndPage(th, page);
-            if(!testHierarchyResult.isEmpty()) {
-                TestHierarchyResult thr = testHierarchyResult.get();
-                JSONObject testHierarchy = new JSONObject();
-                testHierarchy.put("testHierarchyName", th.getName());
-                testHierarchy.put("testHierarchyCode", th.getCode());
-                testHierarchy.put("testHierarchyUrl", th.getUrls());
-                testHierarchy.put("testHierarchyId", th.getId());
-                
-                JSONObject results = new JSONObject();
-                results.put("nb_element_cant_tell", thr.getNbElementCantTell());
-                results.put("nb_cant_tell", thr.getNbCantTell());
-                results.put("nb_element_failed", thr.getNbElementFailed());
-                results.put("nb_element_passed", thr.getNbElementPassed());
-                results.put("nb_element_tested", thr.getNbElementTested());
-                results.put("nb_element_untested", thr.getNbElementUntested());
-                results.put("nb_failed", thr.getNbFailed());
-                results.put("nb_inapplicable", thr.getNbInapplicable());
-                results.put("nb_passed", thr.getNbPassed());
-                results.put("nb_test_cant_tell", thr.getNbTestCantTell());
-                results.put("nb_test_failed", thr.getNbTestFailed());
-                results.put("nb_test_inapplicable", thr.getNbTestInapplicable());
-                results.put("nb_test_passed", thr.getNbTestPassed());
-                results.put("nb_test_untested", thr.getNbTestUntested());
-                results.put("nb_untested", thr.getNbUntested());
-                results.put("status", thr.getStatus());
-
-                testHierarchy.put("testHierarchyResult",results);
-                pageResultObject.append("testHierarchy",testHierarchy);
-            }       
-        }
-    }
+     
 }
